requirements.txt
Flask==1.0.2
flask-restful==0.3.6
Flask-Cors==3.0.6
mysqlclient==1.3.13
SQLAlchemy==1.2.9
Flask-SQLAlchemy==2.3.2 
blinker==1.4
PyJWT==1.6.4

